issue/761: refactor
