--- presentation/application/actor/_actor-sheet.scss
@@ -1,5 +1,0 @@
-textarea.biography {
-  height: 100%;
-  resize: none;
-  overflow-x: hidden;
-}
