+++ presentation/application/item/base/list-item/_base-list-item.scss
