+++ presentation/application/item/fate-card/fate-card-item-sheet-viewmodel.mjs
@@ -1,7 +1,7 @@
 import { ExtenderUtil } from "../../../../common/util/extender-util.mjs";
 import InputNumberSpinnerViewModel from "../../../component/input-number-spinner/input-number-spinner-viewmodel.mjs";
 import BaseItemSheetViewModel from "../base/base-item-sheet-viewmodel.mjs";
-import { DataFieldComponent } from "../base/datafield-component.mjs";
+import { DataFieldComponent } from "../../datafield-component.mjs";
 
 /**
  * @property {TransientFateCard} document
