+++ presentation/application/item/injury/injury-item-sheet-viewmodel.mjs
@@ -2,7 +2,7 @@ import { ExtenderUtil } from "../../../../common/util/extender-util.mjs";
 import CompositeCurrentAndMaximumNumbersViewModel from "../../../component/composite-current-and-maximum-numbers/composite-current-and-maximum-numbers-viewmodel.mjs";
 import InputTextFieldViewModel from "../../../component/input-textfield/input-textfield-viewmodel.mjs";
 import BaseItemSheetViewModel from "../base/base-item-sheet-viewmodel.mjs";
-import { DataFieldComponent } from "../base/datafield-component.mjs";
+import { DataFieldComponent } from "../../datafield-component.mjs";
 
 /**
  * @property {TransientInjury} document
