--- presentation/application/item/item-sheet-subtype.mjs
@@ -1,20 +1,0 @@
-import BaseSheetSubType from "../base-sheet-subtype.mjs";
-
-/**
- * Defines an ItemSheet sub-type, specific to one of the Item document types. 
- * 
- * @abstract Inheritors MUST override:
- * * `template`
- * * `localizedType`
- * * `createViewModel`
- * 
- * Inheritors *may* override: 
- * * `getTitle`
- * * `activateListeners`
- * * `getHeaderButtons`
- * 
- * @extends BaseSheetSubType
- */
-export default class ItemSheetSubType extends BaseSheetSubType {
-  // No custom logic, atm.
-}
