+++ presentation/application/item/momentum-action/momentum-action-item-sheet-viewmodel.mjs
@@ -5,8 +5,8 @@ import InputNumberSpinnerViewModel from "../../../component/input-number-spinner
 import InputRichTextViewModel from "../../../component/input-rich-text/input-rich-text-viewmodel.mjs"
 import InputTextFieldViewModel from "../../../component/input-textfield/input-textfield-viewmodel.mjs"
 import BaseItemSheetViewModel from "../base/base-item-sheet-viewmodel.mjs"
-import { DataFieldComponent } from "../base/datafield-component.mjs"
-import { TemplatedComponent } from "../base/templated-component.mjs"
+import { DataFieldComponent } from "../../datafield-component.mjs"
+import { TemplatedComponent } from "../../templated-component.mjs"
 
 /**
  * @property {TransientMomentumAction} document
