+++ presentation/application/item/skill/skill-item-sheet-viewmodel.mjs
@@ -21,8 +21,8 @@ import DynamicInputDefinition from "../../../dialog/dynamic-input-dialog/dynamic
 import DynamicInputDialog from "../../../dialog/dynamic-input-dialog/dynamic-input-dialog.mjs"
 import ViewModel from "../../../view-model/view-model.mjs"
 import BaseItemSheetViewModel from "../base/base-item-sheet-viewmodel.mjs"
-import { DataFieldComponent } from "../base/datafield-component.mjs"
-import { TemplatedComponent } from "../base/templated-component.mjs"
+import { DataFieldComponent } from "../../datafield-component.mjs"
+import { TemplatedComponent } from "../../templated-component.mjs"
 import ExpertiseTableViewModel from "../expertise/expertise-table-viewmodel.mjs"
 import BaseAttributeListItemViewModel from "./base-attribute/base-attribute-list-item-viewmodel.mjs"
 import SkillPrerequisiteListItemViewModel from "./skill-prerequisite-list-item-viewmodel.mjs"
