+++ presentation/application/templated-component.mjs
@@ -1,4 +1,4 @@
-import { ValidationUtil } from "../../../../common/util/validation-utility.mjs";
+import { ValidationUtil } from "../../common/util/validation-utility.mjs";
 
 /**
  * Represents the definition of a component.
